moved rabbitmq client to pytest-simcore
