+++ packages/service-library/tests/test_rabbitmq.py
@@ -6,7 +6,7 @@
 
 
 import asyncio
-from typing import Any, AsyncIterator, Callable
+from typing import Any, Callable
 from unittest import mock
 
 import aio_pika
@@ -46,32 +46,6 @@ async def test_rabbit_client(rabbit_client_name: str, rabbit_service: RabbitSett
     assert client._connection_pool.is_closed
     assert client._channel_pool
     assert client._channel_pool.is_closed
-
-
-@pytest.fixture
-async def rabbitmq_client(
-    rabbit_service: RabbitSettings,
-) -> AsyncIterator[Callable[[str], RabbitMQClient]]:
-    created_clients = []
-
-    def _creator(client_name: str) -> RabbitMQClient:
-        client = RabbitMQClient(f"pytest_{client_name}", rabbit_service)
-        assert client
-        assert client._connection_pool
-        assert not client._connection_pool.is_closed
-        assert client._channel_pool
-        assert not client._channel_pool.is_closed
-        assert client.client_name == f"pytest_{client_name}"
-        assert client.settings == rabbit_service
-        created_clients.append(client)
-        return client
-
-    yield _creator
-    # cleanup, properly close the clients
-    await asyncio.gather(*(client.close() for client in created_clients))
-    for client in created_clients:
-        assert client._channel_pool
-        assert client._channel_pool.is_closed
 
 
 @pytest.fixture
